requirements.txt
Django==1.7
django-appconf==1.0.1
django-autoslug==1.9.3
django-ckeditor-updated==4.4.4
django-imagekit==3.3
pilkit==1.1.13
Pillow==3.2.0
six==1.10.0
wheel==0.24.0
